pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-ckeditor>=6.7.3",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]
